requirements.txt
certifi==2025.1.31
colorama==0.4.6
coverage==7.6.12
et_xmlfile==2.0.0
idna==3.10
iniconfig==2.0.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1

openai~=1.64.0